refactor: 컨트롤러에서 모델로 이동
