+++ src/main/kotlin/racingcar/model/RacingCarRun.kt
@@ -1,7 +1,5 @@
 package racingcar.model
 
-import racingcar.controller.FinalWinner
-import racingcar.controller.RandomNumberGenerator
 import racingcar.view.InputView
 import racingcar.view.OutputView
